Move action to remote.contribution.ts
